Branch goal: reduce ram usage
Starting reductions bring basic time state usage down to 949 bytes.

+++ source/library/clean/avec.h
@@ -68,8 +68,8 @@ public:
                 delete [] temp;
         }
 private:
-        int curSize;
-        int capacity;
+        uint8_t curSize;
+        uint8_t capacity;
         timeS *data;
 };
